feat: scripts/janken_nash.pyをexamples/07_janken_nash.pyに昇格
エビワラー3すくみ技（マッハパンチ/はやてがえし/きあいパンチ）のNash均衡近似を
学習用サンプルとして整備。優先度・きあいパンチ不発条件に基づく三すくみの関係、
モンテカルロ推定・反復最適反応（fictitious play）・exploitabilityを初心者向けに
解説するdocstring/コメントを追加。examples/READMEに一覧を追記。

スモークテスト（tests/test_examples_smoke.py）で自動実行されることを踏まえ、
研究用の精密設定（GRID_STEP=0.1, EVAL_TRIALS=200, ITERATIONS=30）から
実行時間を抑えた既定値（GRID_STEP=0.25, EVAL_TRIALS=15, ITERATIONS=4、
実行時間約11秒）に変更。精度を求める場合の推奨値はdocstringに明記した。
