+++ examples/README.md
@@ -16,3 +16,7 @@ python examples/01_quickstart.py
 | `04_damage_calculation.py` | `battle.calc_lethal()` による確定数・乱数ダメージ計算 | ダメージ計算ツール開発 |
 | `05_tree_search_ai.py` | `TreeSearchPlayer` を継承した木探索AIとランダム方策の対戦 | AI開発（発展） |
 | `06_bulk_simulation.py` | `Player.battle_against()` による多数回対戦・構成比較 | 戦術研究（構成比較） |
+| `07_janken_nash.py` | 実バトルのモンテカルロ推定 + 反復最適反応によるNash均衡（混合戦略）の近似、`ProcessPoolExecutor` による並列化 | 戦術研究（読み合いの定量分析、発展） |
+
+`07_janken_nash.py` は多数回のバトルシミュレーションを行うため、他のサンプルより実行に
+時間がかかる（手元では十数秒程度）。
